requirements.txt
numpy==1.26.4
scipy==1.11.4
pandas==2.1.4
scikit-learn==1.3.2
psutil==5.9.8
faiss-cpu==1.7.4
hnswlib==0.8.0
requests==2.32.3
matplotlib==3.8.4
PyYAML==6.0.2
